requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
atomicwrites==1.4.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
black==25.1.0
blinker==1.9.0
build==1.2.2.post1
cachelib==0.13.0
cachetools==5.5.2
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
chromadb==1.0.15
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
docstring-to-markdown==0.17
durationpy==0.10
email_validator==2.2.0
eval_type_backport==0.2.2
fastapi==0.116.1
filelock==3.18.0
Flask==2.3.3
Flask-Dance==7.1.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flatbuffers==25.2.10
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
gradio_client==1.10.4
greenlet==3.2.3
grpcio==1.74.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
kubernetes==33.1.0
langchain==0.3.27
langchain-chroma==0.2.5
langchain-community==0.3.27
langchain-core==0.3.72
langchain-ollama==0.3.6
langchain-text-splitters==0.3.9
langsmith==0.4.8
lxml==5.4.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mccabe==0.7.0
mdurl==0.1.2
mistralai==1.9.3
mistune==3.0.2
mmh3==5.2.0
mpmath==1.3.0
msgspec==0.19.0
multidict==6.6.3
mypy_extensions==1.1.0
mysql-connector-python==9.4.0
narwhals==1.46.0
nest-asyncio==1.6.0
numpy==2.3.0
oauthlib==3.3.1
ollama==0.5.1
onnxruntime==1.22.1
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
orjson==3.11.1
overrides==7.7.0
packaging==25.0
pandas==2.3.0
parso==0.8.4
pathspec==0.12.1
patsy==1.0.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
plotly==6.2.0
plotly-express==0.4.1
pluggy==1.6.0
posthog==5.4.0
propcache==0.3.2
protobuf==6.31.1
psutil==5.9.8
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyfiglet==1.0.3
PyGithub==2.3.0
Pygments==2.19.2
PyJWT==2.8.0
pyls-spyder==0.4.0
PyMySQL==1.1.1
PyNaCl==1.5.0
pyparsing==3.2.3
pypdf==5.9.0
PyPika==0.48.9
pyproject_hooks==1.2.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5_sip==12.13.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0
python-dotenv==1.1.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.13.0
pytoolconfig==1.3.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retrying==1.4.0
rich==14.1.0
rope==1.12.0
rpds-py==0.26.0
rsa==4.9.1
Rtree==1.2.0
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
setuptools==78.1.1
shellingham==1.5.4
sip==6.8.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.2
statsmodels==0.14.4
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.4
tqdm==4.67.1
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.9.0
urllib3==2.2.1
URLObject==2.4.3
uvicorn==0.35.0
watchdog==6.0.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==2.3.8
wheel==0.45.1
wrapt==1.17.2
WTForms==3.2.1
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
